test(EuiDataGrid): move EuiDataGrid column cell actions jest tests to a separate file
